microservices/audit_service/requirements.txt
flask==2.3.3
flask-cors==5.0.0
pymongo[srv]==4.10.1
python-dotenv==1.0.1
dnspython==2.7.0
PyJWT==2.8.0